Okay. Filesystem now sets up properly (I think), here were the issues solved:
The first time I called changeDirs() it tried to navigate from the root directory into the root directory due to the way I'm handling having a working directory. The fix was to just skip that first command by changing i to start from 1 instead of 0

The subdirectories were inheriting all of the attributes of the parents becuase I pull everything from an ls command, store it in a list, and then add the things from the list to the current directory. The problem is I forgot to clear said list, so parent attributes were still staying around to be added.

Next step is to traverse tree...